Reduce space complexity of cdtw
Remove the need to store a full matrix for masking purposes; before, used an m*n matrix to mask off the valid warping area (either the projected path for cdtw_fast, or the sakoe-chiba band for cdtw). Now, use a 2*n array to store the upper/lower bounds of the valid warping area for each column of the matrix.